테스트코드 수정

+++ src/main/java/com/wip/bool/music/guitar/service/GuitarCodeService.java
@@ -1,10 +1,14 @@
 package com.wip.bool.music.guitar.service;
 
+import com.wip.bool.exception.excp.AuthorizationException;
 import com.wip.bool.exception.excp.EntityNotFoundException;
 import com.wip.bool.exception.excp.ErrorCode;
 import com.wip.bool.music.guitar.domain.GuitarCode;
 import com.wip.bool.music.guitar.domain.GuitarCodeRepository;
 import com.wip.bool.music.guitar.dto.GuitarCodeDto;
+import com.wip.bool.user.domain.Role;
+import com.wip.bool.user.domain.User;
+import com.wip.bool.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,8 +22,19 @@ public class GuitarCodeService {
 
     private final GuitarCodeRepository guitarCodeRepository;
 
+    private final UserRepository userRepository;
+
     @Transactional
-    public Long saveGuitarCode(GuitarCodeDto.GuitarCodeSaveRequest requestDto) {
+    public Long saveGuitarCode(Long userId, GuitarCodeDto.GuitarCodeSaveRequest requestDto) {
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException(userId, ErrorCode.NOT_FOUND_USER));
+
+        Role role = user.getRole();
+
+        if(role != Role.ROLE_ADMIN) {
+            throw new AuthorizationException();
+        }
 
         int order = guitarCodeRepository.maxOrder();
         GuitarCode guitarCode = GuitarCode.createGuitarCode(requestDto.getCode(), order + 1);
@@ -35,7 +50,16 @@ public class GuitarCodeService {
     }
 
     @Transactional
-    public Long deleteGuitarCode(Long guitarCodeId) {
+    public Long deleteGuitarCode(Long userId, Long guitarCodeId) {
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException(userId, ErrorCode.NOT_FOUND_USER));
+
+        Role role = user.getRole();
+        if(role != Role.ROLE_ADMIN) {
+            throw new AuthorizationException();
+        }
+
         GuitarCode guitarCode = guitarCodeRepository.findById(guitarCodeId)
                     .orElseThrow(() -> new EntityNotFoundException(guitarCodeId, ErrorCode.NOT_FOUND_GUITAR_CODE));
         return guitarCodeRepository.delete(guitarCode);
